Complements the AESNI implementation with CLMUL intrinsics when supported by the CPU.

Before:
```
AE/ChaChaPoly1305                        mean 7.682 μs  ( +- 168.9 ns  )
AE/AES-GCM                               mean 14.32 μs  ( +- 66.21 ns  )
AE/AES-CCM                               mean 21.58 μs  ( +- 96.51 ns  )
```

After (support_pclmuldq disabled):
```
AE/ChaChaPoly1305                        mean 7.728 μs  ( +- 199.7 ns  )
AE/AES-GCM                               mean 14.52 μs  ( +- 94.97 ns  )
AE/AES-CCM                               mean 21.70 μs  ( +- 199.7 ns  )
```

After (support_pclmuldq enabled):
```
AE/ChaChaPoly1305                        mean 7.698 μs  ( +- 170.2 ns  )
AE/AES-GCM                               mean 3.037 μs  ( +- 52.41 ns  )
AE/AES-CCM                               mean 21.65 μs  ( +- 151.6 ns  )
```

For now flag `support_pclmuldq` must still be enabled explicitely.