@kazu-yamamoto 2020-06-12
When will be the next version with `support_pclmuldq` enabled released?