@kazu-yamamoto 2020-06-23
Why does cryptonite-0.27 still disable this flag?