@ocheron 2020-06-28
It's best not to change everything in the same release.
But this can be tried later. Enabling AESNI on Windows too.