@kazu-yamamoto 2020-06-28
@ocheron BTW, do you know that kazuho (my friend) found a way to improve the performance of AES GCM for short objects? 

Blogs are here. Hoping that the translation works well:

http://blog.kazuhooku.com/2020/06/quicaes-gcm-12.html

Do you think that `cryptonite` can support the approach?