@ocheron 2020-07-05
Something like this should be driven by measurements.
Our implementation is very likely to be suboptimal, there can be other factors to improve first.